CIR-127-PortalToGlobalMessages-changes

+++ CIR.Core/Interfaces/Website/IPortal2GlobalConfigurationMessagesRepository.cs
@@ -11,7 +11,7 @@ namespace CIR.Core.Interfaces.Website
 {
 	public interface IPortal2GlobalConfigurationMessagesRepository
 	{
-		Task<IActionResult> GetPortalToGlobalConfigurationMessagesList(int portalId);
+		Task<IActionResult> GetPortalToGlobalConfigurationMessagesList(long portalId);
 		Task<IActionResult> CreateOrUpdatePortalToGlobalConfigurationMessages(List<Portal2GlobalConfigurationMessage> portal2GlobalConfigurationMessages);
 	}
 }
